When no protocol is specified in <port>, listen on IPv6 and IPv4
Ignore-this: 909e5234bf3ce9790b1b3dcf1668249c

If the user doesn't specify a protocol in <port> in the configuration
file, we try to open an IPv6 socket (listening to IPv6 only) if
possible and we open an IPv4 socket anyway. This corresponds to the
net.ipv6.bindv6only=0 behaviour on dual-stack Linux, but should be
more portable and should work with net.ipv6.bindv6only=1 as well.

darcs-hash:20100408115823-0445d-cdfaf9a8051140be031038f147232e2cfc4508f1