Moved transparent element input type to 'a.
Ignore-this: e6da76fd00897a4cf58763391638c0ae
It is now possible to build invalid term like a [a []]
since there is no restriction on the arguments for a. But such
terms will not be valid when passed to top function such as p or body
(in theory).

darcs-hash:20100711183516-99ab4-e5508e8c0276d35d9f8d2d8158d68af26fb93ff6