Eliom: create comet service earlier
Ignore-this: 98d570cfd60658f73d6d8d13afc4f2e

The function checking for cookies to send to the service is in send. If the service is created after, that function can't know that there is a service. The creation is now done in pre_service.

darcs-hash:20110120152553-0cb73-d0eee3595f98f186844433379bc76722d4268e93