Eliom: Make Eliom_appl work with Eliom_output.Any
Ignore-this: 8381b0fc39cd960c382218f4232478b3
Not relying any more on pre_service to set cpi, appl_name and content_only
cpi is created if someone needs it (lazy).
Forms and links must be created by the functions from the right module (created with Eliom_appl),
because they rely on the name of the application set as aparameter of Eliom_appl to decide the kind of link/form.
(was: the name of the application set by the pre_service)

darcs-hash:20110214170245-1b86b-effaacd6e9ac61aa55ec9f727c3ee1fb302c599b