Lazy site values
Ignore-this: 510123336313c4ba1b34350f819d8b15
Those are lazy values available in the context of a site.
It is usefull for registering services before the availability of sp.
Classical lazy values don't work for that because, it would create a
unique service for every site making many problems: the url won't be
correct, the closures would share informations, etc...

The limitations are that these values are never collected: sites always
keeps a reference to them when they are evaluated. It is not a problem
for global services, since there is already a premanent reference to
them, but don't use them for non-persistant datas.

darcs-hash:20110427121617-0cb73-5cb34fbe5d755138f0468d1609349bdb2e593f79