simplification of comet message encoding and fallback service
Ignore-this: be9a017efcb9d1d34f777557a2153f67
Comet messages are now encoded as JSON, it will be easier to extend.
A new kind of message is transmited: Process_closed when the client
process is closed on server side: it is only sent by the new fallback
service. If this service respond, it means that the server cleaned the
process information: the process has been closed.
The client raise the Process_closed exception when this happen.

darcs-hash:20110427135309-0cb73-04672954617f70a4b09e7819b842a173b0f1faf3