Add Eliom_react.S.Down: server -> client signals
Ignore-this: 51a7aa9766a59c633e7a07e0c3714641
The intrest of using this instead of events with React.changes is
that there is no bufferisation: values are sent when they change
and there is a pending request from the client on it.
We can miss some changes if the client does not ask for the value.

darcs-hash:20110505103327-0cb73-2e30ebb9fd34f037240a13b6e500d734097d2cbc