Eliom_appl add an header to tell if the generated page is application content.
Ignore-this: c7349a06d08a4bbb90ca69d4749770ee
This header will be used on client side to react to 404 requests.
It is also used in Eliom_mkreg to choose if the server send a 204 with a
"half redirect" header or if the real page content is sent.

darcs-hash:20110728165151-0cb73-edeb20c853ed79757b72f8e75759169f36d541d5