Making more clear when to use <eliom> and <eliommodule>
Ignore-this: 6be6ce58e34a545b8edcf745aab9b10
It is still possible to use <eliom> to load a module, but a warning is displayed
if it is used several times.
Also failing when using <eliom> to use an eliom extension, because it was not implemented (why?).

darcs-hash:20120822155210-d035b-4b9748aead0e91542bd00dd9e4a3b8952ec940ec