add eliompp, a proprocessor to extract sections
It can be used to extract specific sections. The purpose of this
proprocessor is to keep commentaries in a section. Camlp4 doesn't
include commentaries in the AST (it re-insert them after the
preprocessing of the file) so if you use Camlp4 to keep only
client sections for example, you will also get commentaries from
top-level, server and shared sections. So you can't use it to generate
a proper documentation.