Several history related fixes and improvements.
Do not change the URL on page change for mobile apps: this is not
allowed on iOS due to same origin restrictions.

Add a flag 'replace' to the 'change_page*' functions to replace the
current page in the history.

Fix a bug in my previous commit (the reload function was not properly
stored in the history).