fix server redirection
When encountering a redirection to a service that is only implemented on
the server, the following will occur:
User is on page A and clicks on link to B, which is a redirection to C.
The URL changes to B and then quickly to C when C's content has been
loaded. When the user click on the back button the URL will change to B.
When the user clicks the back button once more the URL will change to A
along with the contents.