configuration option for omitting cookie creation
Adds a configuration option that allows for the omission of the storage
of any persistent information on the server for any requests that
satisfy all of the given conditions. Conditions can be expressed as
PCREs applied to any HTTP header of the request. This can be used to
prevent inundating the cookie table by cookies generated for crawlers
(that usually do not store cookies on their end anyway).